Port all API away from per-entity locale overrides
This makes the entity context the single source of truth for where a
locale is coming from. While this is slightly more work in some corner
cases, in the vast majority of cases it makes reasoning about locale
issues a lot simpler, and makes maintaining the API a lot more
straightforward in future.
In addition to that, it removes a bunch of duplicate locale checking
code and moves it to one central place.