Drop the catchall as-enums unit
We have a lot of enums in different places, and it was not obvious what
belonged there vs other places.
All enums are now kept with their biggest, or most of the time, only
user. This makes it a lot easier to keep track of them and have them
alongside their primary object in the documentation.