Implement environment property for component screenshots
This adds a new "environment" attribute to the "screenshot" element to
specify the environment in which the screenshot was recorded.
It is comprised of a "<env>:<style>" string, where "style" could be a
specific visual style, like "light" or "dark" or even form factor like
"mobile", and "env" is the desktop environment, like "gnome", "plasma"
or even "windows".

Resolves: #390, #410, #304