Don't translate all licenses
License names are part of the legal text as well, so unless there is an
official translation, we may not simply translate them.
This also removes the issue for localization teams to translate e.g.
French license names.
Since SPDX embeds pieces of text that are not part of the license (e.g.
"or later") though, we still allow translation of some licenses.