Rename AsReleases -> AsReleaseList
While "AsReleases" was technically correct and matches the XML, it was
also quite close to "AsRelease" and error-prone to type (which in itself
is not too bad, the compiler usually catches these errors), but it also
makes reading the code really hard in some places - and it's nice to
have code that's also understandable at a glance by the occasional
sleep-deprived PhD student.

It also solves a bunch of semantic issues in the API, like me having a
real hard time to find a plural of "releases" for when releases are
contained in arrays. Having ReleaseList solves all of the issues quite
nicely, at the expense of slightly longer symbol names.