meson.build: Improve stemmer/SnowBall dep search
Don't try to add '/usr/include' to the include directories as it may not exist
on Windows, without checking for its presence.  Likewise, check for
'libstemmer/libstemmer.h' only if '/usr/include' is found.

Also, streamline finding the libstemmer libraries so that we look for
libstemmer.h at the same time as looking for the stemmer library, to ensure
that things are properly set up.