src/*.c: Allow build on clang-cl
clang-cl uses the headers from the Visual Studio and Windows SDK installation,
so make sure that we are including the right headers on Windows for the things
we are using in the code.  In particular, unistd.h is a wrapper header on MinGW
builds which include the various Windows system headers for the things that are
actually done.

Also, for calling umask(), just use S_IREAD | S_IWRITE for the default
permissions on Windows.