as-macros-private.h: Define macros for visibility for clang-cl
On clang-cl (i.e. Visual Studio-like meaning _MSC_VER is defined) builds,
symbols in shared builds have visibility to be hidden by default, and
decorating symbols with GCC's default visibility is not enough to export the
symbol.

Consequently, use the macros in as-desktop-entry.h so that we don't fall into
the situation where clang-cl complains as we are attempting to do a
__declspec(dllexport) in a hidden visibility section.