build: Generate a .def file for Visual Studio-like builds
This way, we can export all the non-static symbols from the appstream DLL that
we are trying to build, without touching the headers too much, by using the
dumpbin tool that comes with every installation of Visual Studio, which clang-cl
uses.

Sadly, this approach does not allow us to effectively filter out those we mark
with the GCC hidden visibility atribute, so this is the best we can do.