Add support for Zstd for on-disk data compression
In measurements, zstd was as fast or even faster than gzip to
decompress, and it provides a small but pretty nice reduction in tarball
size for the larger icon tarballs.
We can't make the switch to zstd for a while, but this change lies the
groundwork to make this possible in future.