utils: Rename the appstream file when re-saving it on install
The as_utils_install_metadata_file() can install the .xml* data
re-saved, with added origin reference. When it does so, and the original
file was compressed, then the file name extension does not reflect
the file content anymore (it's a pure XML after re-saving, not
a compressed XML data). That can confuse the Appstream data readers,
thus rename the file to be .xml, to avoid this problem.