docs: Add workaround for gi-docgen misnaming devhelp files
If --no-namespace-dir and --output-dir are both passed to gi-docgen, the
tool wrongly assumes the output directory root name as namespace name,
which in our case results in wrongly-named devhelp index files (since
our output folder name does not correspond to the namespace name).

We can't easily change the names, as a ton of references to the API docs
online would break. This patch adds a small wrapper that introduces a
--namespace parameter to explicitly set the namespace name to create
correctly-named devhelp files (and in the long run we should address
this upstream in gi-docgen as well).

Resolves: #708