compose: Prefer rendering font icon with "Regular" style
People were supposed to put the font that should show up in the icon as
first-in-list, but people don't seem to read the manual. So we end up
with "Bold Italic" fonts in icons, and similar stuff.

To help font metadata authors, we'll now just always prefer the font of
"Regular" style, and only if there is none fall back to the "first in
list".