compose: Fix possible race condition when fetching pangrams
The code tried to be more clever than was good for it - a GPtrArray is a
complex type. We could do a smarter trick here and use atomic
assignments and locking, but this codepath is not *this* performance
criticial, so using the lock properly is the safest and most
straightforward option.